+++ packages/presets/__tests__/presets.test.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const presets = require('..');
+const assert = require('assert').strict;
+
+assert.strictEqual(presets(), 'Hello from presets');
+console.info('presets tests passed');