+++ playground/my-vue-app/src/main.ts
@@ -1,24 +1,24 @@
-import './assets/main.css'
+import './assets/main.css';
 
-import { createApp } from 'vue'
-import { createPinia } from 'pinia'
-import monitor from "../../../packages/browser/esm";
-import vuePlugin from '../../../packages/vue/esm'
-import performancePlugin from '../../../packages/performance/esm'
-import hashPlugin from '../../../packages/router_hash/esm'
-import historyPlugin from '../../../packages/router_history/esm'
-import xhrPlugin from '../../../packages/xhr/esm'
-import fetchPlugin from '../../../packages/fetch/esm'
+import { createApp } from 'vue';
+import { createPinia } from 'pinia';
+import monitor from '../../../packages/browser/esm';
+// import vuePlugin from '../../../packages/vue/esm'
+// import performancePlugin from '../../../packages/performance/esm'
+// import hashPlugin from '../../../packages/router_hash/esm'
+// import historyPlugin from '../../../packages/router_history/esm'
+// import xhrPlugin from '../../../packages/xhr/esm'
+// import fetchPlugin from '../../../packages/fetch/esm'
+import presets from '../../../packages/presets/esm';
+import App from './App.vue';
+import router from './router';
 
-import App from './App.vue'
-import router from './router'
+const app = createApp(App);
 
-const app = createApp(App)
+app.use(createPinia());
+app.use(router);
 
-app.use(createPinia())
-app.use(router)
-
-app.mount('#app')
+app.mount('#app');
 monitor({
     dsn: {
         reportUrl: 'http://localhost:3000/api/log/upload',
@@ -34,13 +34,14 @@ monitor({
     // },
     debuge: true,
     plugins: [
-        performancePlugin(),
-        vuePlugin({
-            vue: app
-        }),
-        hashPlugin(),
-        historyPlugin(),
-        xhrPlugin(),
-        fetchPlugin()
+        ...presets({ vue: app })
+        // performancePlugin(),
+        // vuePlugin({
+        //     vue: app
+        // }),
+        // hashPlugin(),
+        // historyPlugin(),
+        // xhrPlugin(),
+        // fetchPlugin()
     ]
-})
+});