+++ packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @monitor-sdk/core
 
+## 1.0.5
+
+### Patch Changes
+
+-   功能调整
+-   Updated dependencies
+    -   @monitor-sdk/types@1.0.2
+    -   @monitor-sdk/utils@1.0.3
+
 ## 1.0.4
 
 ### Patch Changes