+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @monitor-sdk/vue
 
+## 1.0.2
+
+### Patch Changes
+
+-   功能调整
+
 ## 1.0.1
 
 ### Patch Changes